Improve Barberias page and add Axios and Geolocation service
The Barberias page has been enhanced with Grid and Box components from material-ui for a better organization and presentation of the BarberiaCards and the MapView. Axios was included to make HTTP requests, mainly to use with the new Geolocation Service, which was also added to get the latitude and longitude for the MapView. The MapView also received improvements, such as the removal of an unused useMap import and the addition of default map coordinates. Changes have also been made in the TestPage to test the new Geolocation Service, and in .css files to adjust the MapView size.