Update user profile page and navbar status
The update includes two main changes. In the user profile file (Perfil.jsx), the page title was slightly modified and a customer profile component (ClientePerfil) was added. This is to provide more informative headers and present user-specific details on the profile page.
Additionally, in the Navbar.jsx file, the isLoggedIn variable was changed from false to true to ensure the visibility of user-specific items in the navbar due to a logged-in status.