Add "Services" page and navigation to it
In the BarberiaCard component, a button was changed to navigate to the new "Services" page when clicked. This page was added to AppRoutes. "Servicios.jsx" and associated components (TopDetailCard, ServicioCard, and BackButton) were created for the new Services page. BackButton has a NavLink to allow navigation back to the previous page. These changes allow users to view individual services offered by a barber shop.