Refactor state handling and improve BackButton styling
The state management related to the cart's services in the Servicios page was moved from component level to custom hooks for more clear and reusable code. This change was implemented by creating a new custom hook, useServiciosState, that manages all the state and functions related to adding and removing services from the cart. This hook is imported and used in the Servicios component.

The useState import and the barberServices array in the Carrito component were removed as they were no longer necessary due to the new state management structure.

The styling for the BackButton component was also improved, adding margins for better spacing around the button.