Update UI and functionality on 'ServicioCard' component
This commit addresses improvements to the 'ServicioCard' and 'Carrito' component. The status of each service in 'ServiciosPage' has been changed from string data ("Disponible", "No disponible") to boolean data (true, false) for better handling. Moreover, new service entries have been added. The typography styling of headers and descriptions in the 'ServicioCard' component and 'Carrito' component are improved to enhance readability. Layout for the 'Carrito' component is wrapped in a Paper component to give it a card-like appearance. The position of the 'Carrito' component is updated to 'sticky' to ensure it stays visible on the screen when the user scrolls.