Add Checkout page and navigation from the cart
Added a Checkout page to navigate to from the Cart. A "Proceed to Checkout" button was added to the Cart, which redirects to new Checkout page when clicked. Also, made changes to the BackButton component, replacing the enclosing Fragment with directly adding the props to the Button component. This simplifies the code and keeps the component more compact and readable. The routing was updated to accommodate the new Checkout page.