Add default barberias and improve user interfaces
Added a default list of barberias in Barberias.jsx for showcasing purposes and to prevent an empty state. Updated ClientePerfil.jsx to include a title for Opinions section. Improved aesthetics of OpinionCard and ClienteCard components by adding elevations, margin, and reformating content. Changed default client name to a more generic one in OpinionCard for a better user understanding.