Update checkout page with new features
Expanded Checkout.jsx by adding new state variables, services data and the necessary function to handle payment. The changes also include the ability to select a date to make an appointment as well as a list of services that the users can book. The changes are made to improve user experience by providing more accurate information to the users before making a payment. In the TopDetailCard component, static values for barber's name and address were replaced with actual data.